add additional arch unit tests
